+++ src/CallableMiddleware.php
@@ -4,10 +4,10 @@ namespace Middlewares\Utils;
 
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use Interop\Http\Middleware\ServerMiddlewareInterface;
-use Interop\Http\Middleware\DelegateInterface;
+use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Interop\Http\ServerMiddleware\DelegateInterface;
 
-class CallableMiddleware implements ServerMiddlewareInterface
+class CallableMiddleware implements MiddlewareInterface
 {
     /**
      * @var callable